refactor(transforms): reorganize into category-based directory structure
Reorganize transforms directory from implementation-based to category-based structure for better maintainability and developer mental models:

- format/ - Code structure & style transforms (prefixing, file-headers, ts-nocheck)
- semantic/ - Semantic enhancement transforms (color-tokens)
- imports/ - Import/dependency transforms (clsx-to-cn)

Preserves all functionality while improving organization and developer experience. All 527 tests continue to pass.
